refactor(core): optimise the scroll to block feature
